+++ model/gurps/export_legacy.go
@@ -464,7 +464,7 @@ func (ex *legacyExporter) emitKey(key string) error {
 	case "PRIMARY_ATTRIBUTE_LOOP_COUNT":
 		count := 0
 		for _, def := range ex.entity.SheetSettings.Attributes.List(true) {
-			if (def.Type != attribute.Pool && def.Type != attribute.PoolRef) && def.Primary() {
+			if (def.Type != attribute.Pool && def.Type != attribute.PoolRef) && def.Primary(ex.entity) {
 				if _, exists := ex.entity.Attributes.Set[def.DefID]; exists {
 					count++
 				}
@@ -476,7 +476,7 @@ func (ex *legacyExporter) emitKey(key string) error {
 	case "SECONDARY_ATTRIBUTE_LOOP_COUNT":
 		count := 0
 		for _, def := range ex.entity.SheetSettings.Attributes.List(true) {
-			if (def.Type != attribute.Pool && def.Type != attribute.PoolRef) && !def.Primary() {
+			if (def.Type != attribute.Pool && def.Type != attribute.PoolRef) && !def.Primary(ex.entity) {
 				if _, exists := ex.entity.Attributes.Set[def.DefID]; exists {
 					count++
 				}
@@ -1168,7 +1168,7 @@ func (ex *legacyExporter) processConditionalModifiersLoop(list []*ConditionalMod
 
 func (ex *legacyExporter) processAttributesLoop(buffer []byte, primary bool) {
 	for _, def := range ex.entity.SheetSettings.Attributes.List(true) {
-		if (def.Type != attribute.Pool && def.Type != attribute.PoolRef) && def.Primary() == primary {
+		if (def.Type != attribute.Pool && def.Type != attribute.PoolRef) && def.Primary(ex.entity) == primary {
 			if attr, ok := ex.entity.Attributes.Set[def.DefID]; ok {
 				ex.processBuffer(buffer, func(key string, _ []byte, index int) int {
 					switch key {